Revert "Footer header"
