--- ControlCash/src/App.js
@@ -1,19 +1,0 @@
-import { StatusBar } from 'expo-status-bar';
-import { StyleSheet, View } from 'react-native';
-import HomeScreen from './Screens/HomeScreens';
-
-
-export default function App() {
-  return (
-      <View style={styles.container}>
-      <HomeScreen/>  
-      <StatusBar style="auto" />
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-});
