--- ControlCash/src/Components/Constants/Coolors.js
@@ -1,21 +1,0 @@
-export const COLORS = {
-  // Fundo geral da tela
-  background: "#F2F5F9",
-
-  // Cores principais
-  azul: "#007BFF", // Azul principal do app
-  azulEscuro: "#0056b3", // Azul para títulos e destaques
-
-  // Texto
-  text: "#1C1C1C",
-  textSecondary: "#6B7280",
-
-  // Cores de feedback
-  sucesso: "#22C55E", // Verde
-  erro: "#EF4444", // Vermelho
-
-  // Elementos
-  cardBackground: "#FFFFFF",
-  borderLight: "rgba(0, 0, 0, 0.05)",
-  white: "#FFFFFF",
-};
